New parser for Åland (#4852)
* new ax parser using iframe

* format data

* fixes + sanity checks

* add myself to contributors

* Rename new parser and archive the old parser
