+++ config/zones/AX.yaml
@@ -13,6 +13,7 @@ capacity:
 contributors:
   - tmslaine
   - systemcatch
+  - VIKTORVAV99
 emissionFactors:
   direct:
     battery discharge:
